# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Tobfa API",
        "contact": {
            "name": "Tobi",
            "email": "tobiaditia549@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "paths": {
...
        "/businesses": {
            "get": {
                "description": "Get Business",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Business"
                ],
                "summary": "Get Business",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.WebResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/web.BusinessResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "Create Business",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Business"
                ],
                "summary": "Create Business",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.BusinessCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.WebResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/web.BusinessResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/businesses/{id}": {
            "get": {
                "description": "Find Business",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Business"
                ],
                "summary": "Find Business",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.WebResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/web.BusinessResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "description": "Update Business",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Business"
                ],
                "summary": "Update Business",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/web.BusinessUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.WebResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/web.BusinessResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Business",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Business"
                ],
                "summary": "Delete Business",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.WebResponse"
                        }
                    }
                }
            }
        },
...
        "/stats/business": {
            "get": {
                "description": "Stats Business",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stat"
                ],
                "summary": "Stats Business",
                "parameters": [
                    {
                        "type": "string",
                        "description": "dateStarted",
                        "name": "dateStarted",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "dateEnded",
                        "name": "dateEnded",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "businessIds, sparated by coma",
                        "name": "businessIds",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "businessTransactionTypeIds, sparated by coma",
                        "name": "businessTransactionTypeIds",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "businessTransactionItemIds, sparated by coma",
                        "name": "businessTransactionItemIds",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/web.WebResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/web.BusinessStatsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "web.Business": {
            "type": "object",
            "properties": {
                "businessTransactionItem": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web.TransactionItemStat"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                },
                "totalBuy": {
                    "type": "integer"
                },
                "totalSell": {
                    "type": "integer"
                }
            }
        },
...
        "web.BusinessCreateRequest": {
            "type": "object",
            "required": [
                "address",
                "businessCategoryId",
                "cityId",
                "countryId",
                "districtId",
                "name",
                "provinceId",
                "userId",
                "villageId"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "businessCategoryId": {
                    "type": "integer"
                },
                "cityId": {
                    "type": "integer"
                },
                "countryId": {
                    "type": "integer"
                },
                "districtId": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                },
                "provinceId": {
                    "type": "integer"
                },
                "userId": {
                    "type": "integer"
                },
                "villageId": {
                    "type": "integer"
                }
            }
        },
        "web.BusinessResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "businessCategoryId": {
                    "type": "integer"
                },
                "cityId": {
                    "type": "integer"
                },
                "countryId": {
                    "type": "integer"
                },
                "districtId": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "provinceId": {
                    "type": "integer"
                },
                "userId": {
                    "type": "integer"
                },
                "villageId": {
                    "type": "integer"
                }
            }
        },
        "web.BusinessStatsResponse": {
            "type": "object",
            "properties": {
                "businessTransactionItemIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "businessTransactionType": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web.TransactionTypeStat"
                    }
                },
                "businessTransactionTypeIds": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "businesses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web.Business"
                    }
                },
                "dateEnded": {
                    "type": "string"
                },
                "dateStarted": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                },
                "totalBuy": {
                    "type": "integer"
                },
                "totalSell": {
                    "type": "integer"
                }
            }
        },
...
        "web.BusinessUpdateRequest": {
            "type": "object",
            "required": [
                "address",
                "businessCategoryId",
                "cityId",
                "countryId",
                "districtId",
                "name",
                "provinceId",
                "userId",
                "villageId"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "businessCategoryId": {
                    "type": "integer"
                },
                "cityId": {
                    "type": "integer"
                },
                "countryId": {
                    "type": "integer"
                },
                "districtId": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                },
                "provinceId": {
                    "type": "integer"
                },
                "userId": {
                    "type": "integer"
                },
                "villageId": {
                    "type": "integer"
                }
            }
        },
...
        "web.TransactionItemStat": {
            "type": "object",
            "properties": {
                "businessTransactionType": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web.TransactionTypeStat"
                    }
                },
                "name": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                },
                "totalBuy": {
                    "type": "integer"
                },
                "totalSell": {
                    "type": "integer"
                }
            }
        },
        "web.TransactionStat": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "web.TransactionTypeStat": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "stats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web.TransactionStat"
                    }
                }
            }
        },
...
        "web.WebResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}